requirements.txt
python-telegram-bot==21.4
praw==7.7.1
pandas==2.2.2
mysql-connector-python==9.0.0